requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
bleach==3.2.1
certifi==2020.6.20
coverage==5.3
dj-database-url==0.5.0
dj-static==0.0.6
Django @ file:///tmp/build/80754af9/django_1605066529526/work
django-crispy-forms==1.10.0
django-debug-toolbar==3.1.1
django-heroku==0.3.1
django-markdownify==0.8.1
elasticsearch==7.10.0
factory-boy==3.1.0
Faker==4.17.1
gunicorn==20.0.4
Markdown==3.3.3
mysqlclient==2.0.1
olefile==0.46
packaging==20.4
Pillow @ file:///C:/ci/pillow_1603823068645/work
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pyparsing==2.4.7
python-dateutil==2.8.1
python-memcached==1.59
pytz==2020.1
selenium==3.141.0
six @ file:///C:/ci/six_1605187374963/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
static3==0.7.0
text-unidecode==1.3
urllib3==1.26.2
webencodings==0.5.1
whitenoise==5.2.0
wincertstore==0.2
